Layout carrinho de compras em andamento!

+++ database/migrations/2021_01_06_105455_create_state_table.php
@@ -7,30 +7,32 @@ use Illuminate\Support\Facades\Schema;
 class CreateStateTable extends Migration
 {
     /**
-     * Run the migrations.
-     *
-     * @return void
-     */
+    * Run the migrations.
+    *
+    * @return void
+    */
     public function up()
     {
-        Schema::create('state', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('country_id')->references('id')->on('country')->onDelete('cascade')->nullable();
-            $table->string('initials', 5)->nullable(); //sigla
-            $table->string('name', 50)->nullable();
-            $table->integer('created_by')->nullable();
-            $table->integer('updated_by')->nullable();
-            $table->integer('deleted_by')->nullable();
-            $table->timestamps();
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('state')) {
+            Schema::create('state', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('country_id')->references('id')->on('country')->onDelete('cascade')->nullable();
+                $table->string('initials', 5)->nullable(); //sigla
+                $table->string('name', 50)->nullable();
+                $table->integer('created_by')->nullable();
+                $table->integer('updated_by')->nullable();
+                $table->integer('deleted_by')->nullable();
+                $table->timestamps();
+                $table->softDeletes();
+            });
+        }
     }
 
     /**
-     * Reverse the migrations.
-     *
-     * @return void
-     */
+    * Reverse the migrations.
+    *
+    * @return void
+    */
     public function down()
     {
         Schema::dropIfExists('state');
